fix move method in Movable interface

+++ BrickInvaders/src/main/java/org/example/Items/Block.java
@@ -29,7 +29,7 @@ public class Block extends Item implements Movable {
     }
 
     @Override
-    public void move(double xCoordinate, double yCoordinate) {
+    public void move(int xCoordinate, int yCoordinate) {
         // does something
         show();
     }
