+++ BrickInvaders/src/main/java/org/example/Items/Boss.java
@@ -34,7 +34,7 @@ public class Boss extends Item implements Movable, Shootable {
     }
 
     @Override
-    public void move(double xCoordinate, double yCoordinate) {
+    public void move(int xCoordinate, int yCoordinate) {
         // does something
         show();
     }
