+++ BrickInvaders/src/main/java/org/example/Items/Shooter.java
@@ -38,7 +38,7 @@ public class Shooter extends Item implements Movable, Shootable {
     }
 
     @Override
-    public void move(double xCoordinate, double yCoordinate) {
+    public void move(int xCoordinate, int yCoordinate) {
         // does something
         show();
     }
